Revert "Func reporte"
