--- .gitignore
@@ -1,65 +1,0 @@
-config/default.py
-__pycache__/
-*.py[cod]
-*$py.class
-
-venv/
-.env/
-.venv/
-ENV/
-env.bak/
-venv.bak/
-
-.env
-*.env
-*.ini
-*.cfg
-*.local
-*.log
-
-*.sqlite3
-*.db
-
-*/__pycache__/
-*/migrations/
-*/__init__.pyc
-
-instance/
-.flaskenv
-
-htmlcov/
-.tox/
-.coverage
-coverage.xml
-.cache
-nosetests.xml
-pytest_cache/
-test-results/
-
-.vscode/
-.idea/
-*.sublime-project
-*.sublime-workspace
-
-*.swp
-*.bak
-*.tmp
-*.orig
-
-_build/
-docs/_build/
-
-pip-wheel-metadata/
-dist/
-build/
-*.egg-info/
-*.egg
-
-MANIFEST
-
-poetry.lock
-Pipfile.lock
-
-.scannerwork/
-config/__pycache__/
-interfaces/controllers/__pycache__/
