+++ domain/models/Notas/asistencia.py
@@ -1,0 +1,8 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+class Asistencia:
+    def __init__(self):
+        self.asistencia_id = None
+        self.estudiante_id = None
+        self.curso_id = None
