+++ domain/models/Notas/calificacion.py
@@ -1,0 +1,13 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+class Calificacion:
+    def __init__(self):
+        self.calificacion_id = None
+        self.estudiante = None
+        self.curso_id = None
+        self.puntaje = None
+        self.Attribute1 = None
+
+    def ponderado(self, peso):
+        pass
