+++ domain/models/Notas/curso.py
@@ -1,0 +1,14 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+class Curso:
+    def __init__(self):
+        self.curso_id = None
+        self.nombre = None
+        self.Attribute1 = None
+
+    def agregar_estudiante(self, estudiante):
+        pass
+
+    def calcular_promedio(self, ):
+        pass
