+++ domain/models/Usuarios/usuario.py
@@ -1,0 +1,12 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+class Usuario:
+    def __init__(self):
+        self.usuario_id = None
+        self.nombres = None
+        self.correo = None
+        self.hash_contraseña = None
+
+    def cambiar_contraseña(self, hash):
+        pass
