+++ domain/models/administrador.py
@@ -1,0 +1,13 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+class Administrador:
+    def __init__(self):
+        self.administrador_id = None
+        self.usuario_id = None
+
+    def crear_usuario(self, usuario):
+        pass
+
+    def eliminar_usuario(self, usuario):
+        pass
