+++ domain/models/anuncio.py
@@ -1,0 +1,11 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+class Anuncio:
+    def __init__(self):
+        self.anuncio_id = None
+        self.curso_id = None
+        self.profesor_id = None
+        self.titulo = None
+        self.contenido = None
+        self.fecha_creado = None
