+++ domain/models/estudiante.py
@@ -1,0 +1,7 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+class Estudiante:
+    def __init__(self):
+        self.estudiante_id = None
+        self.usuario_id = None
