+++ domain/models/iadministradorrepositorio.py
@@ -1,0 +1,18 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+class IAdministradorRepositorio:
+    def __init__(self):
+        pass
+
+    def obtener(self, id):
+        pass
+
+    def agregar(self, admin):
+        pass
+
+    def actualizar(self, admin):
+        pass
+
+    def eliminar(self, id):
+        pass
