+++ domain/models/icursorepositorio.py
@@ -1,0 +1,18 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+class ICursoRepositorio:
+    def __init__(self):
+        pass
+
+    def obtener(self, id):
+        pass
+
+    def agregar(self, curso):
+        pass
+
+    def actualizar(self, curso):
+        pass
+
+    def eliminar(self, id):
+        pass
