+++ domain/models/iprofesorrepositorio.py
@@ -1,0 +1,21 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+class IProfesorRepositorio:
+    def __init__(self):
+        pass
+
+    def obtener(self, id):
+        pass
+
+    def Operation1(self, ):
+        pass
+
+    def agregar(self, profesor):
+        pass
+
+    def actualizar(self, profesor):
+        pass
+
+    def eliminar(self, id):
+        pass
