+++ domain/models/iusuariorepositorio.py
@@ -1,0 +1,18 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+class IUsuarioRepositorio:
+    def __init__(self):
+        pass
+
+    def obtener(self, id):
+        pass
+
+    def agregar(self, usuario):
+        pass
+
+    def actualizar(self, usuario):
+        pass
+
+    def eliminar(self, id):
+        pass
