+++ domain/models/padre.py
@@ -1,0 +1,12 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+class Padre:
+    def __init__(self):
+        self.padre_id = None
+
+    def agregar_estudiante(self, estudiante):
+        pass
+
+    def ver_calificaciones(self, estudiante):
+        pass
