+++ domain/models/profesor.py
@@ -1,0 +1,10 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+class Profesor:
+    def __init__(self):
+        self.profesor_id = None
+        self.usuario_id = None
+
+    def crear_anuncio(self, curso, anuncio):
+        pass
