--- interfaces/repositorio/iadminrepositorio.py
@@ -1,18 +1,0 @@
-#!/usr/bin/python
-# -*- coding: utf-8 -*-
-
-class iadministrador_repositorio:
-    def __init__(self):
-        pass
-
-    def obtener(self, id):
-        pass
-
-    def agregar(self, admin):
-        pass
-
-    def actualizar(self, admin):
-        pass
-
-    def eliminar(self, id):
-        pass
