--- interfaces/repositorio/icalificacion_repositorio.py
@@ -1,6 +1,0 @@
-from abc import ABC, abstractmethod
-
-class ICalificacionRepositorio(ABC):
-    @abstractmethod
-    def agregar(self, calificacion):
-        pass
