--- interfaces/static/css/layout_reporte.css
@@ -1,74 +1,0 @@
-body {
-    margin: 0;
-    font-family: Arial, sans-serif;
-    background-color: #111;
-}
-
-.hero {
-    position: relative;
-    width: 100%;
-    height: 100vh;
-    background-image: url('/static/img/historia.jpg'); /* actualiza nombre */
-    background-size: cover;
-    background-position: center;
-    color: white;
-}
-
-.overlay {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 164px;
-    background-color: rgba(0, 0, 0, 0.5);
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0 40px;
-    box-sizing: border-box;
-}
-
-.overlay .left,
-.overlay .right {
-    display: flex;
-    gap: 20px;
-}
-
-.centered-text {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    text-align: center;
-}
-
-.centered-text h1 {
-    font-size: 48px;
-    margin: 0;
-}
-
-.centered-text p {
-    font-size: 18px;
-}
-
-.grades {
-    max-width: 900px;
-    margin: 40px auto;
-    background: white;
-    border-collapse: collapse;
-}
-
-.grades th,
-.grades td {
-    padding: 16px;
-    border-bottom: 1px solid #ccc;
-    text-align: left;
-}
-
-.grades tr:nth-child(even) {
-    background-color: #f0f0f0;
-}
-
-.grades th {
-    background-color: #eee;
-}
