--- interfaces/static/css/notas_style.css
@@ -1,97 +1,0 @@
-/* Reset básico */
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-body {
-  font-family: "Segoe UI", Roboto, sans-serif;
-  color: #333;
-  line-height: 1.4;
-}
-
-/* HERO HEADER */
-.hero {
-  position: relative;
-  height: 60vh;
-  background: url("../images/header-bg.jpg") center/cover no-repeat;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-.navbar {
-  display: flex;
-  justify-content: space-between;
-  padding: 1rem 2rem;
-}
-
-.logo {
-  color: #fff;
-  text-decoration: none;
-  font-weight: bold;
-}
-
-.nav-links {
-  list-style: none;
-  display: flex;
-  gap: 1.5rem;
-}
-
-.nav-links a {
-  color: #fff;
-  text-decoration: none;
-  font-size: 0.95rem;
-}
-
-.hero-content {
-  text-align: center;
-  color: #fff;
-  margin-bottom: 4rem;
-}
-
-.hero-content h1 {
-  font-size: 3rem;
-  margin-bottom: 0.5rem;
-}
-
-.profesor {
-  font-size: 1.1rem;
-  opacity: 0.8;
-}
-
-/* CONTENIDO */
-.container {
-  max-width: 800px;
-  margin: 2rem auto;
-  padding: 0 1rem;
-}
-
-.grades-table {
-  width: 100%;
-  border-collapse: collapse;
-}
-
-.grades-table thead {
-  background: #f5f5f5;
-}
-
-.grades-table th,
-.grades-table td {
-  padding: 0.75rem 1rem;
-  text-align: left;
-}
-
-.grades-table tbody tr.odd {
-  background: #e0e0e0;
-}
-
-.grades-table tbody tr.even {
-  background: #f0f0f0;
-}
-
-.grades-table th {
-  font-weight: bold;
-  border-bottom: 2px solid #ddd;
-}
