--- interfaces/templates/cursos/cursos.html
@@ -1,80 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <title>Cursos</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
-    <style>
-        .curso-card {
-            height: 120px;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            font-size: 1.2rem;
-            font-weight: bold;
-        }
-    </style>
-</head>
-<body class="p-4">
-
-    <div class="container">
-        <h1 class="mb-4">Gestión de Cursos</h1>
-        
-        {% with messages = get_flashed_messages(with_categories=true) %}
-          {% if messages %}
-            {% for category, message in messages %}
-              <div class="alert alert-{{ category }} alert-dismissible fade show" role="alert">
-                {{ message }}
-                <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
-              </div>
-            {% endfor %}
-          {% endif %}
-        {% endwith %}
-
-        <!-- Botón para mostrar formulario de creación -->
-        <button class="btn btn-primary mb-3" data-bs-toggle="collapse" data-bs-target="#formCrearCurso">
-            + Crear nuevo curso
-        </button>
-
-        <!-- Formulario oculto -->
-        <div class="collapse mb-4" id="formCrearCurso">
-            <form method="POST" action="{{ url_for('curso.crear_curso') }}">
-                <div class="mb-3">
-                    <label for="nombre" class="form-label">Nombre del Curso</label>
-                    <input type="text" class="form-control" id="nombre" name="nombre" required>
-                </div>
-                <button type="submit" class="btn btn-success">Guardar</button>
-            </form>
-        </div>
-
-        {% if cursos %}
-    <div class="row">
-        {% for curso in cursos %}
-        <div class="col-md-4 mb-4">
-            <!-- Toda la tarjeta es un enlace -->
-            <a href="{{ url_for('curso.detalle_curso', curso_id=curso.curso_id) }}" class="text-decoration-none text-dark">
-                <div class="card curso-card shadow-sm text-center d-flex flex-column justify-content-center align-items-center">
-                    <div>{{ curso.nombre }}</div>
-                </div>
-            </a>
-
-            <!-- Botón eliminar debajo del enlace -->
-            <form method="POST"
-                  action="{{ url_for('curso.eliminar_curso', curso_id=curso.curso_id) }}"
-                  class="mt-2"
-                  onsubmit="return confirm('¿Estás seguro de que deseas eliminar este curso?');">
-                <button type="submit" class="btn btn-sm btn-danger">Eliminar</button>
-            </form>
-        </div>
-        {% endfor %}
-    </div>
-{% else %}
-    <div class="alert alert-info">No hay cursos registrados.</div>
-{% endif %}
-
-
-    </div>
-
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-</body>
-</html>
