--- interfaces/templates/cursos/detalle_curso.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <title>Detalles del Curso</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
-</head>
-<body class="p-4">
-    <div class="container">
-        <h1>Curso: {{ curso.nombre }}</h1>
-        <p>ID: {{ curso.id }}</p>
-        <!-- Puedes añadir más detalles del curso aquí -->
-        <a href="{{ url_for('curso.mostrar_cursos') }}" class="btn btn-secondary mt-3">Volver a Cursos</a>
-    </div>
-</body>
-</html>
-
