+++ interfaces/templates/layout.html
@@ -10,7 +10,8 @@
     <title>EDUNET</title>
 </head>
 <body>
-    {% block content %}{% endblock %}
+    {% block body %}
+    {% endblock %}
     <script src="{{ url_for('static', filename='js/layout.js') }}"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.7/dist/js/bootstrap.bundle.min.js" integrity="sha384-ndDqU0Gzau9qJ1lfW4pNLlhNTkCfHzAVBReH9diLvGRem5+R9g2FzA8ZGN954O5Q" crossorigin="anonymous"></script>
 </body>
