+++ interfaces/templates/principal/index.html
@@ -11,7 +11,7 @@
                 <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
                 <div class="navbar-nav ms-auto d-flex align-items-center gap-2">
                     <a class="nav-link active" aria-current="page" href="#">Home</a>
-                     <a class="nav-link" href="{{ url_for('curso.mostrar_cursos') }}">Cursos</a>
+                    <a class="nav-link" href="#">Features</a>
                     <button type="button" class="btn btn-dark">Iniciar Sesión</button>
                 </div>
                 </div>
