--- interfaces/templates/reporte/reporte.html
@@ -1,44 +1,0 @@
-{% extends "reporte/layout_reporte.html" %}
-
-{% block title %}Notas - Historia{% endblock %}
-
-{% block body %}
-<div class="hero">
-    <div class="overlay">
-        <div class="left">
-            <span>Cuenta Padre</span>
-        </div>
-        <div class="right">
-            <span>Anuncios</span>
-            <span>Notas</span>
-        </div>
-    </div>
-    <div class="centered-text">
-        <h1>Historia</h1>
-        <p>Profesor Juan Perez</p>
-    </div>
-</div>
-
-<table class="grades">
-    <thead>
-        <tr>
-            <th>Asignación</th>
-            <th>Nota</th>
-        </tr>
-    </thead>
-    <tbody>
-        <tr>
-            <td>Tarea 1 - Lectura Imperio Incaico</td>
-            <td>15</td>
-        </tr>
-        <tr>
-            <td>Tarea 2 - Virreyes del Perú</td>
-            <td>10</td>
-        </tr>
-        <tr>
-            <td>Examen Unidad 1</td>
-            <td>13</td>
-        </tr>
-    </tbody>
-</table>
-{% endblock %}
