--- service/mysql/base.py
@@ -1,4 +1,0 @@
-# filepath: domain/models/base.py
-from sqlalchemy.orm import declarative_base
-
-Base = declarative_base()
