--- service/mysql/calificacion_repositorio_impl.py
@@ -1,13 +1,0 @@
-from interfaces.repositorio.icalificacion_repositorio import ICalificacionRepositorio
-from models import calificacion
-
-class CalificacionRepositorioImpl(ICalificacionRepositorio):
-    def __init__(self, session):
-        self.session = session
-
-    def agregar(self, calificacion):
-        self.session.add(calificacion)
-        self.session.commit()
-
-    def obtener_por_estudiante(self, estudiante_id):
-        return self.session.query(Calificacion).filter_by(estudiante_id=estudiante_id).all()
